Don't try to fully destroy a comment that has children
It violates a foreign key constraint, which is the technical
representation of the product reality that we probably shouldn't allow
totally deleting comments that have children/replies, because that would
effectively make the children totally irrelevant / not displayed.
Instead of totally destroying it, if a Comment has one or more children,
we will "blank it out".

+++ app/controllers/api/comments_controller.rb
@@ -21,9 +21,13 @@ class Api::CommentsController < Api::BaseController
 
     authorize(comment)
 
-    comment.destroy!
-
-    head :ok
+    if comment.children.exists?
+      comment.update!(content: '[deleted]', user: nil)
+      render json: comment
+    else
+      comment.destroy!
+      head :no_content
+    end
   end
 
   def index
